refactor: Input 컴포넌트 내부 속성 constants로 분리
