+++ src/constants/atoms/avatar.ts
